chore: refactor embedding
