+++ aperag/embed/test_image_embedding.py
@@ -6,53 +6,12 @@ import requests
 from langchain_core.prompts import PromptTemplate
 
 from aperag.query.query import QueryWithEmbedding, get_packed_answer
-from aperag.readers.base_embedding import get_embedding_model
-from aperag.readers.local_path_embedding import LocalPathEmbedding
+from aperag.embed.base_embedding import get_embedding_model
+from aperag.embed.local_path_embedding import LocalPathEmbedding
 from aperag.utils.date import elapsed_time
 from aperag.vectorstore.connector import VectorStoreConnectorAdaptor
 
 # CFG = Config()
-"""
-REFINE_TEMPLATE = (
-    "### System:\n"
-    "You are an artificial assistant that gives facts based answers.\n"
-    "You strive to answer concisely.\n"
-    "When you're done responding, create and append a terse review to the answer.\n"
-    "In your review, you review the response to fact check it and point out any inaccuracies.\n"
-    "Be analytical and critical in your review, and very importantly, don't repeat parts of your answer.\n"
-    "### Human:\n"
-    "The original question is as follows: {query_str}\n"
-    "We have provided an existing answer: {existing_answer}\n"
-    "We have the opportunity to refine the existing answer "
-    "(only if needed) with some more context below.\n"
-    "------------\n"
-    "{context_msg}\n"
-    "------------\n"
-    "Given the new context, refine the original answer to better \n"
-    "answer the question. \n"
-    "If the context isn't useful, return the original answer.\n"
-    "### Assistant :\n"
-)"""
-
-REFINE_TEMPLATE = (
-    "### System:\n"
-    "You are an artificial assistant that gives facts based answers.\n"
-    "You strive to refine and synthesize the provided existing answers and context message.\n"
-    "In your review, you review the response to fact check it and point out any inaccuracies.\n"
-    "Be analytical and critical in your review, and very importantly, don't repeat parts of your answer.\n"
-    "### Human:\n"
-    "The original question is as follows: {query_str}\n"
-    "We have provided an existing answer: {existing_answer}\n"
-    "We have the opportunity to refine the existing answer "
-    "(only if needed) with some more context below.\n"
-    "------------\n"
-    "{context_msg}\n"
-    "------------\n"
-    "Given the new context, refine the original answer to better \n"
-    "answer the question. \n"
-    "If the context isn't useful, return the original answer.\n"
-    "### Assistant:\n"
-)
 
 VICUNA_REFINE_TEMPLATE = (
     "### Human:\n"
@@ -95,7 +54,7 @@ def test_local_path_embedding_query(query: str, collection_name: str):
         "timeout": 1000,
     }
     adaptor = VectorStoreConnectorAdaptor("qdrant", ctx)
-    embedding, vector_size = get_embedding_model("huggingface")
+    embedding, _ = get_embedding_model("huggingface")
     vector = embedding.get_query_embedding(query)
     query_embedding = QueryWithEmbedding(query=query, top_k=3, embedding=vector)
 
@@ -120,21 +79,6 @@ def get_streaming_response(llm_server: str, input: str):
 
 def stream(llm_server: str, input: str):
     try:
-        # Initialize ncurses
-        # screen = curses.initscr()
-        # curses.noecho()
-        # curses.cbreak()
-        # screen.keypad(True)
-
-        # Clear screen
-        # screen.clear()
-
-        # Print initial message
-        # screen.addstr(0, 0, "Hello, ncurses world!")
-
-        # Move cursor
-        # screen.move(5, 5)
-
         s = requests.Session()
         r = s.post("%s/generate_stream" % llm_server, json=input, stream=True)
         buffer = ""
@@ -150,12 +94,8 @@ def stream(llm_server: str, input: str):
             if "}" in c:
                 j = json.loads(buffer)
                 buffer = ""
-                # screen.addstr(0, 0, f"{j['text']}")
-                # screen.refresh()
+                print(j["text"])
     finally:
-        # Close ncurses
-        # curses.nocbreak(); screen.keypad(0); curses.echo()
-        # curses.endwin()
         pass
 
 
@@ -170,7 +110,7 @@ def test_local_llm_qa(query: str, collection_name: str):
     }
     adaptor = VectorStoreConnectorAdaptor("qdrant", ctx)
 
-    embedding, vector_size = get_embedding_model("huggingface")
+    embedding, _ = get_embedding_model("huggingface")
     vector = embedding.get_query_embedding(query)
     query_embedding = QueryWithEmbedding(query=query, top_k=10, embedding=vector)
 
@@ -182,7 +122,7 @@ def test_local_llm_qa(query: str, collection_name: str):
         limit=query_embedding.top_k,
     )
 
-    answer_text = get_packed_answer(hits.results, 3500)
+    answer_text = get_packed_answer(hits.results, 1900)
 
     prompt = PromptTemplate.from_template(VICUNA_REFINE_TEMPLATE)
     prompt_str = prompt.format(query_str=query, existing_answer=answer_text)
@@ -192,7 +132,7 @@ def test_local_llm_qa(query: str, collection_name: str):
     input = {
         "prompt": prompt_str,
         "temperature": 0,
-        "max_new_tokens": 4096,
+        "max_new_tokens": 2048,
         "model": "gptj-6b",
         "stop": "### Assistant:",
     }
@@ -201,7 +141,7 @@ def test_local_llm_qa(query: str, collection_name: str):
     print(len(prompt_str))
 
     start = time.time()
-    llm_server = "http://47.98.164.173:8000"
+    llm_server = "http://47.96.65.19:8000"
 
     # for line in get_streaming_response(llm_server, input):
     # run stream in a real terminal is ok, for emulator terminal, it is not well adapated
@@ -216,8 +156,10 @@ def start():
     # test_local_path_embedding_query("what is data lake")
     # test_local_llm_qa("what is data lake", "paper")
     os.environ.setdefault("PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION", "python")
-    # test_local_path_embedding(path="/Users/slc/pics/", collection_name="pics")
-    test_local_llm_qa("what is etcd balancer", "pics")
+    test_local_path_embedding(
+        path="../resources/datasets/menu_pics", collection_name="pics2"
+    )
+    test_local_llm_qa("any menu about shrimp?", "pics2")
 
 
 if __name__ == "__main__":
