+++ aperag/views/admin.py
@@ -21,7 +21,7 @@ from ninja import Router
 from config import settings
 from aperag.context.context import ContextManager
 from aperag.db.ops import query_collection_without_user
-from aperag.readers.base_embedding import get_embedding_model
+from aperag.embed.base_embedding import get_embedding_model
 from aperag.rank.reranker import rerank
 from aperag.source.utils import async_run
 from aperag.utils.request import get_user
